requirements.txt
slackclient
python-dotenv
